requirements.txt
# Core Framework
Flask==3.0.3
Werkzeug==3.0.1
flask-restful==0.3.10
flask-cors==4.0.0

# Database and ORM
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
alembic==1.13.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
pymysql==1.1.0

# Authentication and Security
Flask-JWT-Extended==4.6.0
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# Task Queue and Caching
celery==5.3.6
redis==5.0.1
Flask-Caching==2.1.0

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.4.0
marshmallow==3.21.1
Flask-Marshmallow==1.2.0

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==22.7.0

# Development Tools
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
pytz==2024.1
python-dateutil==2.8.2
gunicorn==21.2.0

# Monitoring and Logging
prometheus-client==0.19.0
flask-prometheus-metrics==1.0.0
structlog==24.1.0

# Analysis and ML Tools
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Environment Management
pip-tools==7.3.0


oauthlib==3.2.2
requests-oauthlib==1.3.1